new file;
This file is only needed by those running the development (CVS)
version of phpESP.

There is one critical issue which will probably effect you.
When I first added the ability to store more than one
!other question, I inadvertantly stored the wrong ID.
So ... The current values in the 'choice_id' column of
'answers_other' do not correctly correspond to the *real*
choice ID. They do however distinguish between the other
'!other's for that question. The problem is that the
results display will label them wrong.

This file will do some intelligent guessing and calculation
and output a SQL script that *should* fix all the problems.
It does not however do any of the updates itself, it is up to
the user to look at the script and sanity check it, then run it
by hand.

If you are upgrading and have questions please email me. -James


git-svn-id: svn://svn.code.sf.net/p/phpesp/code/trunk@57 b3e5a24c-6043-0410-a8b3-ed4e85eb3ebe